#125 fix clang compilation -Wshorten-64-to-32

+++ dynawo/sources/API/TL/TLTimeline.cpp
@@ -55,7 +55,7 @@ Timeline::eventEquals(const Event& left, const Event& right) const {
 
 int
 Timeline::getSizeEvents() {
-  return events_.size();
+  return static_cast<int>(events_.size());
 }
 
 Timeline::event_const_iterator
